add missing tests, remove obsolete ones
